+++ backend/app/controllers/OdinsonController.scala
@@ -699,6 +699,15 @@ class OdinsonController @Inject() (
       val depsVocabSize = {
         loadVocabulary.terms.toSet.size
       }
+      val fields = MultiFields.getFields(extractorEngine.indexReader)
+      val fieldNames = fields.iterator.asScala.toList
+      val storedFields =
+        if (extractorEngine.numDocs < 1) {
+          Nil
+        } else {
+          val firstDoc = extractorEngine.doc(0)
+          firstDoc.iterator.asScala.map(_.name).toList
+        }
       val tokenFields = extractorEngine.dataGatherer.storedFields
       val allFields = MultiFields.getFields(extractorEngine.indexReader)
       val allFieldNames = allFields.iterator.asScala.toList
@@ -709,7 +718,8 @@ class OdinsonController @Inject() (
         "corpus" -> corpusDir,
         "distinctDependencyRelations" -> depsVocabSize,
         "tokenFields" -> tokenFields,
-        "docFields" -> docFields
+        "docFields" -> docFields,
+        "storedFields" -> storedFields
       )
       json.format(pretty)
     }
@@ -968,10 +978,14 @@ class OdinsonController @Inject() (
   ) = Action.async {
     Future {
       try {
-        val od: OdinsonDocument = loadParentDocByDocumentId(documentId)
-        val json: JsValue = Json.parse(od.toJson)("metadata")
+        val odinsonDocument: OdinsonDocument = loadParentDocByDocumentId(documentId)
+        val json: JsValue = Json.parse(odinsonDocument.toJson)("metadata")
         json.format(pretty)
       } catch {
+        case _: NullPointerException =>
+          InternalServerError(
+            "This search index does not have document filenames saved as stored fields, so metadata cannot be retrieved."
+          )
         case NonFatal(e) =>
           val stackTrace = ExceptionUtils.getStackTrace(e)
           val json = Json.toJson(Json.obj("error" -> stackTrace))
@@ -987,12 +1001,17 @@ class OdinsonController @Inject() (
     Future {
       try {
         val extractorEngine: ExtractorEngine = newEngine()
-        val luceneDoc: LuceneDocument = extractorEngine.indexReader.document(sentenceId)
+
+        val luceneDoc: LuceneDocument = extractorEngine.doc(sentenceId)
         val documentId = luceneDoc.getValues(docIdField).head
-        val od: OdinsonDocument = loadParentDocByDocumentId(documentId)
-        val json: JsValue = Json.parse(od.toJson)("metadata")
+        val odinsonDocument: OdinsonDocument = loadParentDocByDocumentId(documentId)
+        val json: JsValue = Json.parse(odinsonDocument.toJson)("metadata")
         json.format(pretty)
       } catch {
+        case _: NullPointerException =>
+          InternalServerError(
+            "This search index does not have document filenames saved as stored fields, so metadata cannot be retrieved."
+          )
         case NonFatal(e) =>
           val stackTrace = ExceptionUtils.getStackTrace(e)
           val json = Json.toJson(Json.obj("error" -> stackTrace))
@@ -1008,12 +1027,16 @@ class OdinsonController @Inject() (
     Future {
       try {
         val extractorEngine: ExtractorEngine = newEngine()
-        val luceneDoc: LuceneDocument = extractorEngine.indexReader.document(sentenceId)
+        val luceneDoc: LuceneDocument = extractorEngine.doc(sentenceId)
         val documentId = luceneDoc.getValues(docIdField).head
-        val od: OdinsonDocument = loadParentDocByDocumentId(documentId)
-        val json: JsValue = Json.parse(od.toJson)
+        val odinsonDocument: OdinsonDocument = loadParentDocByDocumentId(documentId)
+        val json: JsValue = Json.parse(odinsonDocument.toJson)
         json.format(pretty)
       } catch {
+        case _: NullPointerException =>
+          InternalServerError(
+            "This search index does not have document filenames saved as stored fields, so the parent document cannot be retrieved."
+          )
         case NonFatal(e) =>
           val stackTrace = ExceptionUtils.getStackTrace(e)
           val json = Json.toJson(Json.obj("error" -> stackTrace))
@@ -1032,6 +1055,10 @@ class OdinsonController @Inject() (
         val json: JsValue = Json.parse(odinsonDoc.toJson)
         json.format(pretty)
       } catch {
+        case _: NullPointerException =>
+          InternalServerError(
+            "This search index does not have document filenames saved as stored fields, so the parent document cannot be retrieved."
+          )
         case NonFatal(e) =>
           val stackTrace = ExceptionUtils.getStackTrace(e)
           val json = Json.toJson(Json.obj("error" -> stackTrace))
@@ -1147,8 +1174,6 @@ class OdinsonController @Inject() (
 
   def loadParentDocByDocumentId(documentId: String): OdinsonDocument = {
     val extractorEngine: ExtractorEngine = newEngine()
-    //val doc = extractorEngine.indexSearcher.doc(odinsonScoreDoc.doc)
-    //val fileName = doc.getField(fileName).stringValue
     // lucene doc containing metadata
     val parentDoc: LuceneDocument = extractorEngine.getParentDoc(documentId)
     val odinsonDocFile = new File(docsDir, parentDoc.getField(parentDocFileName).stringValue)