+++ CHANGES.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Added tags-vocabulary endpoint to API for index-specific part-of-speech tags.
+- Added tests for metadata and parent API calls in backend.
 ### Changed
+- Parent document filenames are stored by default.
 
 ## [0.4.0] - 2021-06-07
 ### Added